src/main/java/com/vinhosales/vinhos/dto/CarrinhoDTO.java
package com.vinhosales.vinhos.dto;

import java.time.LocalDateTime;
import java.util.List;

public class CarrinhoDTO {

    private String id;
    private List<ProdutoDTO> produtos;
    private LocalDateTime dataCriacao;
    private LocalDateTime dataAlteracao;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public List<ProdutoDTO> getProdutos() {
        return produtos;
    }

    public void setProdutos(List<ProdutoDTO> produtos) {
        this.produtos = produtos;
    }

    public LocalDateTime getDataCriacao() {
        return dataCriacao;
    }

    public void setDataCriacao(LocalDateTime dataCriacao) {
        this.dataCriacao = dataCriacao;
    }

    public LocalDateTime getDataAlteracao() {
        return dataAlteracao;
    }

    public void setDataAlteracao(LocalDateTime dataAlteracao) {
        this.dataAlteracao = dataAlteracao;
    }
}
